+++ src/test/java/l9g/app/drivemount/config/DrivemountPropertiesTest.java
@@ -29,7 +29,7 @@ class DrivemountPropertiesTest
     return new DrivemountProperties(
       new DrivemountProperties.Keycloak("https://idp.example.org", "r", "c",
         "s"),
-      "email", "fallbackdomain", Duration.ofSeconds(2), shares);
+      "email", "fallbackdomain", Duration.ofSeconds(2), null, shares);
   }
 
   @Test